Rejig paper-nms dependencies to not say -SNAPSHOT
At some point, I should write some scripts or release procedures for this, but until then we'll just pretend that this magically works, because this is what we've deployed to Maven Central. This will almost certainly not work with GitHub Actions. Feel free to contact me and I'll give you the low-down on how to get this working.